pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai_agent_framework"
version = "0.1.0"
description = "A flexible framework for building AI agents using simple, composable patterns"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.8"
dependencies = [
    "anthropic>=0.21.0",
    "aiohttp>=3.8.0",
    "pyyaml>=6.0",
    "fastapi>=0.103.0",
    "uvicorn>=0.23.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.4.0",
    "flake8>=6.0.0",
]
web = [
    "fastapi>=0.103.0",
    "uvicorn>=0.23.0",
]
data = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/ai_agent_framework"
"Bug Tracker" = "https://github.com/yourusername/ai_agent_framework/issues"

[project.scripts]
ai-agent = "ai_agent_framework.main:cli_entry_point"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true